Changed nps to string, and then realized g4 cannot handle more than int32 nps, so we will need to revert, but I made some other minor QOL improvements at the same time so I'm committing anyway

+++ src/PaleoSimOutputManager.cc
@@ -83,7 +83,7 @@ void PaleoSimOutputManager::CreateOutputFileAndTrees() {
   ////////////////////
   fHeaderTree = new TTree("headerTree","Run meta information");
 
-  int nps = fMessenger.GetNPS();
+  long long nps = fMessenger.GetNPS();
   char sourceType[256] = "";
   strcpy(sourceType, fMessenger.GetSourceType().substr(0, std::min<size_t>(255, fMessenger.GetSourceType().length())).c_str());
